requirements-dev.txt
-c requirements.txt
-r requirements.txt
coverage==5.1
django-coverage-plugin==1.8.0
django-debug-toolbar==3.2
django-graphiql-debug-toolbar==0.1.4
black==19.10b0
pylint==2.7.2
pylint-django==2.4.2
watchdog==0.9.0
pytest===5.3.2
pytest-django===3.9.0
factory-boy===2.12.0
isort===5.7.0